Re-enable preprocessing for doxygen
`__host__` and `__device__` function attributes are shown in the docs.

Re-enable preprocessing for doxygen, adds cuda / hip attributes
to `cpp_id_attributes` in sphinx this fixes it erroring out on these
attributes.
The rocprim namespace is added for all function name references in the
rst docs, they were not working previously because the namespace macros
were not expanded by doxygen.